requirements.txt
openai-whisper==20231117
pydub==0.25.1
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3